# add quick logic for jobs to be passed in the uri path