--- a/labels_test.go
+++ b/labels_test.go
@@ -22,9 +22,10 @@ func TestFormatLabels(t *testing.T) {
 	}
 	a.formatLabels(m, "test")
 
-	assert.Equal(t, &dto.LabelPair{Name: strPtr("thing1"), Value: strPtr("value1")}, m.Label[0])
-	assert.Equal(t, &dto.LabelPair{Name: strPtr("thing2"), Value: strPtr("value2")}, m.Label[1])
-	assert.Len(t, m.Label, 2)
+	assert.Equal(t, &dto.LabelPair{Name: strPtr("job"), Value: strPtr("test")}, m.Label[0])
+	assert.Equal(t, &dto.LabelPair{Name: strPtr("thing1"), Value: strPtr("value1")}, m.Label[1])
+	assert.Equal(t, &dto.LabelPair{Name: strPtr("thing2"), Value: strPtr("value2")}, m.Label[2])
+	assert.Len(t, m.Label, 3)
 
 }